archieve/immersed_boundary_endpoint/requirements.txt
numpy>=1.20.0
matplotlib>=3.5.0
scipy>=1.8.0
pysdf>=0.1.9
PyMCubes>=0.1.2
scikit-image>=0.19.0 